@@ -36,19 +36,19 @@
     },
     "private": true,
     "dependencies": {
-        "@angular/animations": "^11.2.4",
+        "@angular/animations": "12.0.2",
         "@angular/cdk": "^11.2.4",
-        "@angular/common": "^11.2.4",
-        "@angular/compiler": "^11.2.4",
-        "@angular/core": "^11.2.4",
+        "@angular/common": "12.0.2",
+        "@angular/compiler": "12.0.2",
+        "@angular/core": "12.0.2",
         "@angular/flex-layout": "^11.0.0-beta.33",
-        "@angular/forms": "^11.2.4",
+        "@angular/forms": "12.0.2",
         "@angular/material": "^11.2.4",
-        "@angular/platform-browser": "^11.2.4",
-        "@angular/platform-browser-dynamic": "^11.2.4",
-        "@angular/router": "^11.2.4",
+        "@angular/platform-browser": "12.0.2",
+        "@angular/platform-browser-dynamic": "12.0.2",
+        "@angular/router": "12.0.2",
         "@auth0/angular-jwt": "^5.0.2",
-        "@ckeditor/ckeditor5-angular": "^2.0.1",
+        "@ctrl/ngx-emoji-mart": "^6.0.0",
         "@ctrl/ngx-rightclick": "^4.0.0",
         "@datorama/akita": "6.1.3",
         "@datorama/akita-ng-entity-service": "^6.0.0",
@@ -137,27 +148,27 @@
         "@types/lodash": "^4.14.168",
         "@types/mongoose": "5.7.36",
         "@types/mongoose-paginate-v2": "1.3.8",
-        "@types/node": "12.12.38",
+        "@types/node": "14.14.33",
         "@types/sanitize-html": "^1.27.1",
         "@types/voca": "^1.4.0",
-        "@typescript-eslint/eslint-plugin": "4.3.0",
-        "@typescript-eslint/parser": "4.3.0",
+        "@typescript-eslint/eslint-plugin": "4.19.0",
+        "@typescript-eslint/parser": "4.19.0",
         "cross-env": "^7.0.3",
         "cypress": "^6.0.1",
-        "dotenv": "6.2.0",
+        "dotenv": "8.2.0",
         "electron": "^11.3.0",
-        "eslint": "7.10.0",
-        "eslint-config-prettier": "6.0.0",
+        "eslint": "7.22.0",
+        "eslint-config-prettier": "8.1.0",
         "eslint-plugin-cypress": "^2.10.3",
         "exitzero": "^1.0.1",
         "jest": "26.2.2",
-        "jest-preset-angular": "8.3.2",
+        "jest-preset-angular": "8.4.0",
         "ngx-tailwind": "2.1.0",
         "nx-electron": "^11.0.0",
         "prettier": "2.2.1",
         "tailwindcss": "^2.1.2",
-        "ts-jest": "26.4.0",
+        "ts-jest": "26.5.5",
         "ts-node": "~9.1.1",
-        "typescript": "~4.0.3"
+        "typescript": "4.2.4"
     }
 }
