@@ -1,16 +1,9 @@
-import { NgModule } from '@angular/core';
+import { NgModule, SecurityContext } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { HttpClientModule, HTTP_INTERCEPTORS } from '@angular/common/http';
+import { HTTP_INTERCEPTORS, HttpClientModule } from '@angular/common/http';
 import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { FormsModule, ReactiveFormsModule } from '@angular/forms';
 import { MAT_SNACK_BAR_DEFAULT_OPTIONS } from '@angular/material/snack-bar';
-import { NgxsModule } from '@ngxs/store';
-import { NgxsReduxDevtoolsPluginModule } from '@ngxs/devtools-plugin';
-import { NgxsStoragePluginModule } from '@ngxs/storage-plugin';
-import { NgxsLoggerPluginModule } from '@ngxs/logger-plugin';
-import { NgxsDispatchPluginModule } from '@ngxs-labs/dispatch-decorator';
-import { NgxsSelectSnapshotModule } from '@ngxs-labs/select-snapshot';
-import { NgxsRouterPluginModule } from '@ngxs/router-plugin';
 import { FileUploadModule } from 'ng2-file-upload';
 import { CookieModule } from 'ngx-cookie';
 import { Ng2FittextModule } from 'ng2-fittext';
@@ -37,7 +30,6 @@ import { PortfolioPages } from './pages/portfolio';
 import { MessagesPages } from './pages/messages';
 import { ContentViewPages } from './pages/content-views';
 import { ErrorPages } from './pages/errors';
-import { MigrationPages } from './pages/migration';
 
 /* Components */
 import { AppComponent } from './app.component';
@@ -47,7 +39,6 @@ import { ContentComponents } from './components/content';
 import { UserComponents } from './components/user';
 import { UserSettingsComponents } from './components/user/settings';
 import { CollectionsComponents } from './components/content/collections';
-import { CommentsComponents } from './components/content/comments';
 import { FriendsComponents } from './components/user/friends';
 import { HistoryComponents } from './components/user/history';
 
@@ -55,16 +46,11 @@ import { HistoryComponents } from './components/user/history';
 import { AppRoutingModule } from './app-routing.module';
 import { IconsModule } from '@dragonfish/client/icons';
 import { PipesModule } from '@dragonfish/client/pipes';
-import { EditorModule } from '@dragonfish/client/editor';
 import { UiModule } from '@dragonfish/client/ui';
 import { MaterialModule } from '@dragonfish/client/material';
 import { AlertsModule } from '@dragonfish/client/alerts';
-import { MyStuffModule } from '@dragonfish/client/my-stuff';
-import { DashboardModule } from '@dragonfish/client/dashboard';
 import { ClientServicesModule } from '@dragonfish/client/services';
-
-/* State */
-import { RepositoryModule } from '@dragonfish/client/repository';
+import { CommentsModule } from '@dragonfish/client/comments';
 
 /* Util */
 import { environment } from '../environments/environment';
@@ -79,7 +65,6 @@ import { AuthInterceptor } from '@dragonfish/client/repository/session/services'
         ...UserComponents,
         ...UserSettingsComponents,
         ...CollectionsComponents,
-        ...CommentsComponents,
         ...FriendsComponents,
         ...HistoryComponents,
         ...HomePages,
@@ -90,7 +75,6 @@ import { AuthInterceptor } from '@dragonfish/client/repository/session/services'
         ...MessagesPages,
         ...ContentViewPages,
         ...ErrorPages,
-        ...MigrationPages,
     ],
     imports: [
         BrowserModule,
@@ -105,9 +89,7 @@ import { AuthInterceptor } from '@dragonfish/client/repository/session/services'
         UiModule,
         MaterialModule,
         AlertsModule,
-        EditorModule,
-        MyStuffModule,
-        DashboardModule,
+        CommentsModule,
         ClientServicesModule,
         Ng2FittextModule,
         NgxPaginationModule,
@@ -120,35 +102,14 @@ import { AuthInterceptor } from '@dragonfish/client/repository/session/services'
         ClickOutsideModule,
         ContextMenuModule,
         CookieModule.forRoot(),
-        MarkdownModule.forRoot(),
-        RepositoryModule,
-        NgxsModule.forRoot(
-            [], {
-                developmentMode: !environment.production,
-                selectorOptions: { suppressErrors: false, injectContainerState: false },
-            }),
-        NgxsStoragePluginModule.forRoot({
-            key: [
-                'auth.token',
-                'user',
-                'global',
-                'myStuff.currContent',
-                'myStuff.currContentWordCount',
-                'myStuff.sections.currSection'
-            ],
-        }),
-        NgxsReduxDevtoolsPluginModule.forRoot(),
-        NgxsRouterPluginModule.forRoot(),
-        NgxsDispatchPluginModule.forRoot(),
-        NgxsSelectSnapshotModule.forRoot(),
-        NgxsLoggerPluginModule.forRoot({ disabled: environment.production }),
+        MarkdownModule.forRoot({ sanitize: SecurityContext.NONE }),
         environment.production ? [] : AkitaNgDevtools.forRoot(),
         AkitaNgRouterStoreModule,
     ],
     providers: [
         { provide: HTTP_INTERCEPTORS, useClass: AuthInterceptor, multi: true },
-        { provide: MAT_SNACK_BAR_DEFAULT_OPTIONS, useValue: { duration: 2500 }},
-        { provide: NG_ENTITY_SERVICE_CONFIG, useValue: { baseUrl: 'https://jsonplaceholder.typicode.com' }},
+        { provide: MAT_SNACK_BAR_DEFAULT_OPTIONS, useValue: { duration: 2500 } },
+        { provide: NG_ENTITY_SERVICE_CONFIG, useValue: { baseUrl: 'https://jsonplaceholder.typicode.com' } },
     ],
     bootstrap: [AppComponent],
 })
