@@ -16,7 +16,7 @@ import {
 import { ApiTags } from '@nestjs/swagger';
 import { FileInterceptor } from '@nestjs/platform-express';
 import { OptionalAuthGuard, RolesGuard } from '../../guards';
-import { ContentFilter, ContentKind, FormType, PubChange } from '@dragonfish/shared/models/content';
+import { ContentFilter, ContentKind, FormType, PubChange, PubContent } from '@dragonfish/shared/models/content';
 import { Roles } from '@dragonfish/shared/models/users';
 import { isNullOrUndefined } from '@dragonfish/shared/functions';
 import { User } from '@dragonfish/api/utilities/decorators';
@@ -24,12 +24,15 @@ import { JwtPayload } from '@dragonfish/shared/models/auth';
 import { DragonfishTags } from '@dragonfish/shared/models/util';
 import { IContent } from '../../shared/content';
 import { IImages } from '../../shared/images';
+import { CommentStore } from '@dragonfish/api/database/comments/stores';
+import { CommentKind } from '@dragonfish/shared/models/comments';
 
 @Controller('content')
 export class ContentController {
     constructor(
         @Inject('IContent') private readonly content: IContent,
-        @Inject('IImages') private readonly images: IImages
+        @Inject('IImages') private readonly images: IImages,
+        private readonly comments: CommentStore,
     ) {}
 
     @ApiTags(DragonfishTags.Content)
@@ -49,15 +52,20 @@ export class ContentController {
     async fetchOnePublished(
         @User() user: JwtPayload,
         @Query('contentId') contentId: string,
-        @Query('kind') kind: ContentKind
-    ) {
+        @Query('kind') kind: ContentKind,
+        @Query('page') page: number,
+    ): Promise<PubContent> {
         if (isNullOrUndefined(contentId) && isNullOrUndefined(kind)) {
             throw new BadRequestException(`You must include the content ID and the content kind in your request.`);
         }
 
         const [content, ratings] = await this.content.fetchOnePublished(contentId, kind, user);
-
-        return { content: content, ratings: ratings };
+        const comments = await this.comments.fetch(content._id, CommentKind.ContentComment, page);
+        return {
+            content: content,
+            ratings: ratings,
+            comments: comments,
+        };
     }
 
     @ApiTags(DragonfishTags.Content)
@@ -73,7 +81,7 @@ export class ContentController {
         @Query('filter') filter: ContentFilter,
         @Query('pageNum') pageNum: number,
         @Query('userId') userId: string,
-        @Query('kind') kind: ContentKind[]
+        @Query('kind') kind: ContentKind[],
     ) {
         if (isNullOrUndefined(pageNum) && isNullOrUndefined(kind)) {
             throw new BadRequestException(`You must include both the page number and content kind in your request.`);
@@ -100,7 +108,7 @@ export class ContentController {
         @User() user: JwtPayload,
         @Query('contentId') contentId: string,
         @Query('kind') kind: ContentKind,
-        @Body() formInfo: FormType
+        @Body() formInfo: FormType,
     ) {
         if (isNullOrUndefined(contentId) || isNullOrUndefined(kind)) {
             throw new BadRequestException(`You must include both the content ID and content kind with this request.`);
@@ -138,11 +146,11 @@ export class ContentController {
     async uploadProseCoverArt(
         @UploadedFile() coverArtImage: any,
         @User() user: JwtPayload,
-        @Param('proseId') proseId: string
+        @Param('proseId') proseId: string,
     ) {
         const coverArtUrl = await this.images.upload(coverArtImage, proseId, 'coverart');
         const coverArt = `${process.env.IMAGES_HOSTNAME}/coverart/${coverArtUrl.substr(
-            coverArtUrl.lastIndexOf('/') + 1
+            coverArtUrl.lastIndexOf('/') + 1,
         )}`;
         return await this.content.updateCoverArt(user, proseId, ContentKind.ProseContent, coverArt);
     }
@@ -154,11 +162,11 @@ export class ContentController {
     async uploadPoetryCoverArt(
         @UploadedFile() coverArtImage: any,
         @User() user: JwtPayload,
-        @Param('poetryId') poetryId: string
+        @Param('poetryId') poetryId: string,
     ) {
         const coverArtUrl = await this.images.upload(coverArtImage, poetryId, 'coverart');
         const coverArt = `${process.env.IMAGES_HOSTNAME}/coverart/${coverArtUrl.substr(
-            coverArtUrl.lastIndexOf('/') + 1
+            coverArtUrl.lastIndexOf('/') + 1,
         )}`;
         return await this.content.updateCoverArt(user, poetryId, ContentKind.PoetryContent, coverArt);
     }
