@@ -34,7 +34,7 @@ export class AppComponent implements OnInit, AfterViewInit {
     ) {}
 
     ngOnInit(): void {
-        this.appQuery.theme$.pipe(untilDestroyed(this)).subscribe(theme => {
+        this.appQuery.theme$.pipe(untilDestroyed(this)).subscribe((theme) => {
             const body = document.getElementsByTagName('body')[0];
             const currTheme = body.classList.item(0);
             const html = document.getElementsByTagName('html')[0];
@@ -57,7 +57,7 @@ export class AppComponent implements OnInit, AfterViewInit {
             }
         });
 
-        this.router.events.pipe(untilDestroyed(this), delay(300)).subscribe(event => {
+        this.router.events.pipe(untilDestroyed(this), delay(300)).subscribe((event) => {
             if (this.sidenav) {
                 if (event instanceof NavigationStart) {
                     this.sidenavService.close();
@@ -86,7 +86,7 @@ export class AppComponent implements OnInit, AfterViewInit {
     }
 
     ngAfterViewInit() {
-        this.sessionQuery.currentUser$.pipe(untilDestroyed(this), delay(300)).subscribe(user => {
+        this.sessionQuery.currentUser$.pipe(untilDestroyed(this), delay(300)).subscribe((user) => {
             if (user !== null || !this.electron.isElectronApp) {
                 this.sidenavService.setSidenav(this.sidenav);
             }
