@@ -1,19 +1,13 @@
 import { Component } from '@angular/core';
-import { Select } from '@ngxs/store';
-import { Observable } from 'rxjs';
-import { Collection } from '@dragonfish/shared/models/collections';
-import { CollectionsState } from '@dragonfish/client/repository/collections';
 import { CollectionFormComponent } from '../collection-form/collection-form.component';
 import { MatDialog } from '@angular/material/dialog';
 
 @Component({
     selector: 'dragonfish-collections-toolbar',
     templateUrl: './collections-toolbar.component.html',
-    styleUrls: ['./collections-toolbar.component.scss']
+    styleUrls: ['./collections-toolbar.component.scss'],
 })
 export class CollectionsToolbarComponent {
-    @Select(CollectionsState.currCollection) currCollection$: Observable<Collection>;
-
     constructor(private dialog: MatDialog) {}
 
     /**
