@@ -1,6 +1,5 @@
 import { CollectionsToolbarComponent } from './collections-toolbar/collections-toolbar.component';
 import { CollectionCardComponent } from './collection-card/collection-card.component';
 import { CollectionFormComponent } from './collection-form/collection-form.component';
-import { AddToCollectionComponent } from './add-to-collection/add-to-collection.component';
 
-export const CollectionsComponents = [CollectionsToolbarComponent, CollectionCardComponent, CollectionFormComponent, AddToCollectionComponent];
+export const CollectionsComponents = [CollectionsToolbarComponent, CollectionCardComponent, CollectionFormComponent];
