@@ -1,6 +1,4 @@
 import { PostRatingComponent } from './post-rating/post-rating.component';
-import { ContentRatingComponent } from './content-rating/content-rating.component';
 import { ContentBoxComponent } from './content-box/content-box.component';
-import { NewsRatingComponent } from './news-rating/news-rating.component';
 
-export const ContentComponents = [PostRatingComponent, ContentRatingComponent, ContentBoxComponent, NewsRatingComponent];
+export const ContentComponents = [PostRatingComponent, ContentBoxComponent];
