@@ -1,10 +1,8 @@
 import { Component, OnInit } from '@angular/core';
-import { Select } from '@ngxs/store';
-import { Observable } from 'rxjs';
-import { HistoryService, HistoryState, HistoryStateModel } from '@dragonfish/client/repository/history';
 import { PopupModel } from '@dragonfish/shared/models/util';
 import { PopupComponent } from '@dragonfish/client/ui';
 import { MatDialog } from '@angular/material/dialog';
+import { ReadingHistoryQuery, ReadingHistoryService } from '@dragonfish/client/repository/reading-history';
 
 @Component({
     selector: 'dragonfish-history-feed',
@@ -12,15 +10,17 @@ import { MatDialog } from '@angular/material/dialog';
     styleUrls: ['./history-feed.component.scss'],
 })
 export class HistoryFeedComponent implements OnInit {
-    @Select(HistoryState) historyState$: Observable<HistoryStateModel>;
-
-    constructor(public historyService: HistoryService, private dialog: MatDialog) {}
+    constructor(
+        public historyService: ReadingHistoryService,
+        private dialog: MatDialog,
+        public historyQuery: ReadingHistoryQuery,
+    ) {}
 
     ngOnInit(): void {
-        this.historyService.fetch();
+        this.historyService.fetch().subscribe();
     }
 
-    askDelete(docIds: string[]) {
+    askDelete() {
         const alertData: PopupModel = {
             message: 'Are you sure you want to delete this? This action is irreversible.',
             confirm: true,
@@ -28,7 +28,7 @@ export class HistoryFeedComponent implements OnInit {
         const dialogRef = this.dialog.open(PopupComponent, { data: alertData });
         dialogRef.afterClosed().subscribe((wantsToDelete: boolean) => {
             if (wantsToDelete) {
-                this.historyService.delete(docIds);
+                this.historyService.delete(this.historyQuery.currentIds as string[]).subscribe();
             }
         });
     }
