@@ -6,15 +6,30 @@ import { MatDialogRef } from '@angular/material/dialog';
 import { SessionQuery } from '@dragonfish/client/repository/session';
 import { UserService } from '@dragonfish/client/repository/session/services';
 import { untilDestroyed } from '@ngneat/until-destroy';
+import { animate, state, style, transition, trigger } from '@angular/animations';
+import { isNullOrUndefined } from '@dragonfish/shared/functions';
 
 @Component({
     selector: 'dragonfish-upload-avatar',
     templateUrl: './upload-avatar.component.html',
-    styleUrls: ['./upload-avatar.component.scss']
+    styleUrls: ['./upload-avatar.component.scss'],
+    animations: [
+        // the fade-in/fade-out animation.
+        trigger('simpleFadeAnimation', [
+            // the "in" style determines the "resting" state of the element when it is visible.
+            state('in', style({ opacity: 1 })),
+
+            // fade in when created. this could also be written as transition('void => *')
+            transition(':enter', [style({ opacity: 0 }), animate(600)]),
+
+            // fade out when destroyed. this could also be written as transition('void => *')
+            transition(':leave', animate(600, style({ opacity: 0 }))),
+        ]),
+    ],
 })
 export class UploadAvatarComponent implements OnInit {
     token: string;
-    imageChangedEvent: Event;
+    imageChangedEvent: any;
     croppedImage: any = '';
 
     fileToReturn: File;
@@ -25,7 +40,7 @@ export class UploadAvatarComponent implements OnInit {
     uploader: FileUploader = new FileUploader({
         url: '/api/user/upload-avatar',
         itemAlias: 'avatar',
-        headers: []
+        headers: [],
     });
 
     constructor(
@@ -36,26 +51,35 @@ export class UploadAvatarComponent implements OnInit {
     ) {}
 
     ngOnInit() {
-        this.sessionQuery.token$.pipe(untilDestroyed(this)).subscribe(value => {
+        this.sessionQuery.token$.pipe(untilDestroyed(this)).subscribe((value) => {
             this.token = value;
         });
     }
 
-    fileChangeEvent(event: any): void {
-        this.imageChangedEvent = event;
+    fileChangeEvent(event: Event | File[]): void {
+        if (isNullOrUndefined(event[0]) !== true) {
+            const fileEvent = event as File[];
+            if (fileEvent.length > 1) {
+                this.alerts.error(`You're only allowed to upload one image!`);
+                return;
+            }
+            this.imageChangedEvent = {
+                target: {
+                    accept: 'image/png, image/jpeg, image/jpg',
+                    files: event as File[],
+                },
+            };
+            this.showCropper = true;
+        } else {
+            this.imageChangedEvent = event;
+            this.showCropper = true;
+        }
     }
 
     imageCropped(event: ImageCroppedEvent): File {
         this.croppedImage = event.base64;
-
         this.fileToReturn = this.base64ToFile(event.base64, 'avatar');
-
-        console.log(this.fileToReturn);
         return this.fileToReturn;
-    }
-
-    imageLoaded(): void {
-        this.showCropper = true;
     }
 
     loadImageFailed(): void {
@@ -86,11 +110,14 @@ export class UploadAvatarComponent implements OnInit {
         this.uploader.clearQueue();
         this.uploader.addToQueue([this.fileToReturn]);
 
-        this.user.changeAvatar(this.uploader).subscribe(() => {
-            this.uploading = false;
-            this.dialogRef.close();
-        }, () =>{
-            this.uploading = false;
-        });
+        this.user.changeAvatar(this.uploader).subscribe(
+            () => {
+                this.uploading = false;
+                this.dialogRef.close();
+            },
+            () => {
+                this.uploading = false;
+            },
+        );
     }
 }
