@@ -1,25 +1,29 @@
 import { Injectable } from '@angular/core';
-import { Resolve, ActivatedRouteSnapshot, RouterStateSnapshot, UrlSegment } from '@angular/router';
+import { Resolve, ActivatedRouteSnapshot } from '@angular/router';
 import { Observable } from 'rxjs';
-import { ContentKind, ContentModel } from '@dragonfish/shared/models/content';
+import { ContentKind, PubContent } from '@dragonfish/shared/models/content';
 import { ContentViewService } from '@dragonfish/client/repository/content-view';
-import { RatingsModel } from '@dragonfish/shared/models/ratings';
 
 @Injectable()
-export class ContentViewResolver implements Resolve<Observable<{ content: ContentModel, ratings: RatingsModel }>> {
+export class ContentViewResolver implements Resolve<Observable<PubContent>> {
+    private pageNum = 1;
+
     constructor(private viewService: ContentViewService) {}
 
-    resolve(route: ActivatedRouteSnapshot, routerState: RouterStateSnapshot) {
+    resolve(route: ActivatedRouteSnapshot) {
         const contentId = route.paramMap.get('contentId');
+        if (route.queryParamMap.has('page')) {
+            this.pageNum = +route.queryParamMap.get('page');
+        }
 
         if (route.url[0].path === 'prose') {
-            return this.viewService.fetchContent(contentId, ContentKind.ProseContent);
+            return this.viewService.fetchContent(contentId, ContentKind.ProseContent, this.pageNum);
         } else if (route.url[0].path === 'poetry') {
-            return this.viewService.fetchContent(contentId, ContentKind.PoetryContent);
+            return this.viewService.fetchContent(contentId, ContentKind.PoetryContent, this.pageNum);
         } else if (route.url[0].path === 'news') {
-            return this.viewService.fetchContent(contentId, ContentKind.NewsContent);
+            return this.viewService.fetchContent(contentId, ContentKind.NewsContent, this.pageNum);
         } else if (route.url[0].path === 'post') {
-            return this.viewService.fetchContent(contentId, ContentKind.BlogContent);
+            return this.viewService.fetchContent(contentId, ContentKind.BlogContent, this.pageNum);
         }
     }
 }
