@@ -1,26 +1,27 @@
 import { Component, Input } from '@angular/core';
 import { Location } from '@angular/common';
-import { Select, Store } from '@ngxs/store';
-import { Observable } from 'rxjs';
-import { ApprovalQueueState, AQNamespace } from '../../../shared/approval-queue';
 import { ApprovalQueue } from '@dragonfish/shared/models/approval-queue';
 import { ContentModel } from '@dragonfish/shared/models/content';
 import { Decision } from '@dragonfish/shared/models/contrib';
 import { UserInfo } from '@dragonfish/shared/models/users';
-import { Navigate } from '@ngxs/router-plugin';
 import { ActivatedRoute, Router } from '@angular/router';
+import { ApprovalQueueService, ApprovalQueueQuery } from '@dragonfish/client/repository/dashboard/approval-queue';
 
 @Component({
     selector: 'dragonfish-approval-queue-toolbar',
     templateUrl: './approval-queue-toolbar.component.html',
-    styleUrls: ['./approval-queue-toolbar.component.scss']
+    styleUrls: ['./approval-queue-toolbar.component.scss'],
 })
 export class ApprovalQueueToolbarComponent {
-    @Select(ApprovalQueueState.selectedDoc) selectedDoc$: Observable<ApprovalQueue>;
     @Input() route: ActivatedRoute;
     @Input() pageNum: number;
 
-    constructor(private location: Location, private store: Store, private router: Router) {}
+    constructor(
+        private location: Location,
+        private router: Router,
+        private queueService: ApprovalQueueService,
+        public queueQuery: ApprovalQueueQuery,
+    ) {}
 
     forceRefresh() {
         // eventually, none of this will be necessary and items will be updated in-place.
@@ -43,9 +44,9 @@ export class ApprovalQueueToolbarComponent {
             workId: thisWork._id,
             authorId: thisWorksAuthor._id,
         };
-        this.store
-            .dispatch([new AQNamespace.ApproveWork(decision), new Navigate(['/dashboard/approval-queue'])])
-            .subscribe();
+        this.queueService.approveWork(decision).subscribe(() => {
+            this.router.navigate(['/dashboard/approval-queue']);
+        });
     }
 
     rejectWork(doc: ApprovalQueue) {
@@ -56,8 +57,8 @@ export class ApprovalQueueToolbarComponent {
             workId: thisWork._id,
             authorId: thisWorksAuthor._id,
         };
-        this.store
-            .dispatch([new AQNamespace.RejectWork(decision), new Navigate(['/dashboard/approval-queue'])])
-            .subscribe();
+        this.queueService.rejectWork(decision).subscribe(() => {
+            this.router.navigate(['/dashboard/approval-queue']);
+        });
     }
 }
