@@ -1,6 +1,67 @@
 import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 import { FeatherModule } from 'angular-feather';
+import { RemixIconModule } from 'angular-remix-icon';
+
+import {
+    RiBold,
+    RiItalic,
+    RiUnderline,
+    RiStrikethrough,
+    RiLink,
+    RiLinkUnlink,
+    RiAlignLeft,
+    RiAlignRight,
+    RiAlignCenter,
+    RiAlignJustify,
+    RiDoubleQuotesL,
+    RiText,
+    RiH1,
+    RiH2,
+    RiH3,
+    RiSeparator,
+    RiListUnordered,
+    RiArrowGoForwardLine,
+    RiArrowGoBackLine,
+    RiImageAddLine,
+    RiFilmLine,
+    RiEmotionHappyLine,
+    RiCloseLine,
+    RiCheckLine,
+    RiHardDrive2Line,
+    RiChatSmile3Line,
+    RiHome5Line,
+    RiCompass3Line,
+    RiGroupLine,
+    RiDashboardLine,
+    RiMore2Line,
+    RiLoginCircleLine,
+    RiLogoutCircleLine,
+    RiHistoryLine,
+    RiNotification3Line,
+    RiSettings5Line,
+    RiUserLine,
+    RiUserAddLine,
+    RiHeartLine,
+    RiDislikeLine,
+    RiChat3Line,
+    RiEyeLine,
+    RiArrowRightSLine,
+    RiArrowLeftSLine,
+    RiHeartFill,
+    RiDislikeFill,
+    RiBookMarkLine,
+    RiFlagLine,
+    RiQuillPenLine,
+    RiBookOpenLine,
+    RiThumbUpLine,
+    RiBarChartLine,
+    RiArrowDownSLine,
+    RiPenNibLine,
+    RiCalendar2Line,
+    RiSearchLine,
+    RiCupLine,
+} from 'angular-remix-icon';
 
 import {
     Rss,
@@ -102,6 +163,13 @@ import {
     Link,
     ExternalLink,
     Gift,
+    Underline,
+    Film,
+    Smile,
+    CornerUpRight,
+    AlignLeft,
+    AlignRight,
+    AlignCenter,
 } from 'angular-feather/icons';
 
 const icons = {
@@ -204,11 +272,78 @@ const icons = {
     Link,
     ExternalLink,
     Gift,
+    Underline,
+    Film,
+    Smile,
+    CornerUpRight,
+    AlignLeft,
+    AlignRight,
+    AlignCenter,
+};
+
+const remixIcons = {
+    RiBold,
+    RiItalic,
+    RiUnderline,
+    RiStrikethrough,
+    RiLink,
+    RiLinkUnlink,
+    RiAlignLeft,
+    RiAlignCenter,
+    RiAlignRight,
+    RiAlignJustify,
+    RiDoubleQuotesL,
+    RiText,
+    RiH1,
+    RiH2,
+    RiH3,
+    RiListUnordered,
+    RiSeparator,
+    RiArrowGoForwardLine,
+    RiArrowGoBackLine,
+    RiImageAddLine,
+    RiFilmLine,
+    RiEmotionHappyLine,
+    RiCloseLine,
+    RiCheckLine,
+    RiHardDrive2Line,
+    RiChatSmile3Line,
+    RiHome5Line,
+    RiCompass3Line,
+    RiGroupLine,
+    RiDashboardLine,
+    RiMore2Line,
+    RiLoginCircleLine,
+    RiLogoutCircleLine,
+    RiHistoryLine,
+    RiNotification3Line,
+    RiSettings5Line,
+    RiUserLine,
+    RiUserAddLine,
+    RiHeartLine,
+    RiDislikeLine,
+    RiChat3Line,
+    RiEyeLine,
+    RiArrowRightSLine,
+    RiArrowLeftSLine,
+    RiHeartFill,
+    RiDislikeFill,
+    RiBookMarkLine,
+    RiFlagLine,
+    RiQuillPenLine,
+    RiBookOpenLine,
+    RiThumbUpLine,
+    RiBarChartLine,
+    RiArrowDownSLine,
+    RiPenNibLine,
+    RiCalendar2Line,
+    RiSearchLine,
+    RiCupLine,
 };
 
 @NgModule({
     declarations: [],
-    imports: [CommonModule, FeatherModule.pick(icons)],
-    exports: [FeatherModule],
+    imports: [CommonModule, FeatherModule.pick(icons), RemixIconModule.configure(remixIcons)],
+    exports: [FeatherModule, RemixIconModule],
 })
 export class IconsModule {}
