@@ -1,12 +1,14 @@
 import { Injectable } from '@angular/core';
 import { Query } from '@datorama/akita';
 import { ContentViewState } from './content-view.state';
-import { ContentViewStore } from '@dragonfish/client/repository/content-view/content-view.store';
+import { ContentViewStore } from './content-view.store';
 
 @Injectable({ providedIn: 'root' })
 export class ContentViewQuery extends Query<ContentViewState> {
     public state$ = this.select();
     public currContent$ = this.select('currContent');
+    public currComments$ = this.select('currPageComments');
+    public loading$ = this.selectLoading();
     public currSection$ = this.select('currSection');
     public ratingsDoc$ = this.select('ratingsDoc');
 
@@ -24,6 +26,10 @@ export class ContentViewQuery extends Query<ContentViewState> {
 
     public get currRating() {
         return this.getValue().currRating;
+    }
+
+    public get currPage() {
+        return this.getValue().currPage;
     }
 
     public get likes() {
